Extract printCompilationMessage method to another file
